Move fixed set of figures into docs/images dir
Moved reference figures from root dir into doc/images/figures/reference/
and current set of calculated figures from <language>/calcs/ into
doc/images/figures/<language>.

This means the displayed figures in the README won't get over-written
when the calculations are re-run, and it's a conscious decision to
update the figures show in the README with a new set.

It's also in preparation for Ford documentation where thes will be re-
used.